[radiation_damping] Fixed cast error detected on Win7 MinGW 4.9.0
[ 89%] Building CXX object hdb_interpolators/unit_tests/CMakeFiles/hdb_interpolators_tests.dir/src/hdb_test.cpp.obj
c:/Jenkins/jobs/Simulator_TEST/workspace/code/hdb_interpolators/unit_tests/src/hdb_test.cpp: In function 'std::function<double(double)> get_interpolated_Br()':
c:/Jenkins/jobs/Simulator_TEST/workspace/code/hdb_interpolators/unit_tests/src/hdb_test.cpp:27:107: error: conversion to 'double' from 'size_t {aka long long unsigned int}' may alter its value [-Werror=conversion]
     auto omegas = builder.build_exponential_intervals_reversed(omega_min, omega_med, (size_t)std::floor(N*r+0.5));
                                                                                                           ^
compilation terminated due to -Wfatal-errors.
cc1plus.exe: all warnings being treated as errors
make[3]: *** [hdb_interpolators/unit_tests/CMakeFiles/hdb_interpolators_tests.dir/src/hdb_test.cpp.obj] Error 1
make[2]: *** [hdb_interpolators/unit_tests/CMakeFiles/hdb_interpolators_tests.dir/all] Error 2
make[1]: *** [CMakeFiles/run_all_tests.dir/rule] Error 2
make: *** [run_all_tests] Error 2
Build step 'Exécuter un script shell' marked build as failure

+++ code/hdb_interpolators/unit_tests/src/hdb_test.cpp
@@ -22,8 +22,8 @@ std::function<double(double)> get_interpolated_Br()
     RadiationDampingBuilder builder(TypeOfInterpolation::SPLINES, TypeOfQuadrature::SIMPSON);
     const double r = 0.5;
     const double omega_med = builder.find_r_bound(test_data::analytical_Br,omega_min,omega_max, r);
-    auto omegas = builder.build_exponential_intervals_reversed(omega_min, omega_med, (size_t)std::floor(N*r+0.5));
-    const auto omegas2 = builder.build_exponential_intervals(omega_med, omega_max, (size_t)std::floor(N*(1-r)+0.5));
+    auto omegas = builder.build_exponential_intervals_reversed(omega_min, omega_med, (size_t)std::floor((double)N*r+0.5));
+    const auto omegas2 = builder.build_exponential_intervals(omega_med, omega_max, (size_t)std::floor((double)N*(1-r)+0.5));
     omegas.insert(omegas.end(), omegas2.begin()+1,omegas2.end());
     for (auto omega:omegas) vBr.push_back(test_data::analytical_Br(omega));
     return builder.build_interpolator(omegas,vBr);
